change balance and wagerResult types to double

+++ app/src/main/java/com/example/betbracket/players/Player.kt
@@ -2,9 +2,9 @@ package com.example.betbracket.players
 
 data class Player(
     var name: String,
-    var balance: Int,
+    var balance: Double,
     var image: String,
-    var lastWagerResult: Int = 0
+    var lastWagerResult: Double = 0.0
     ){
 
 }
